v0.0.1-preview update

+++ app/admin/model/dvmodel/host.go
@@ -6,7 +6,7 @@ type Host struct {
 	model.Base
 	Name     string `json:"name" gorm:"type:varchar(50);not null;index:idx_name;comment:名称"`
 	Ip       string `json:"ip" gorm:"type:varchar(50);not null;index:idx_ip,unique;comment:IP"`
-	Port     int16  `json:"port" gorm:"type:int;not null;comment:ssh端口"`
+	Port     int32  `json:"port" gorm:"type:int;not null;comment:ssh端口"`
 	User     string `json:"user" gorm:"type:varchar(50);not null;comment:ssh用户名"`
 	Password string `json:"password" gorm:"type:varchar(255);not null;comment:ssh密码"`
 	SortNum  int32  `json:"sortNum" gorm:"type:int;default:0;comment:顺序"`
